optimizes Tekmovanje

+++ Programiranje/Java/Tekmovanje/Testi/Nakljucno/Tekm_63250060.java
@@ -7,11 +7,11 @@ public class Tekm_63250060 {
 	public static void main(String[] args) {
 
 		Scanner s = new Scanner(System.in);
-		byte tip = s.nextByte();
-		byte n = s.nextByte();
-		byte k = s.nextByte();
+		int tip = s.nextInt();
+		int n = s.nextInt();
+		int k = s.nextInt();
 		BigInteger m = new BigInteger(s.next());
-		byte[] result = tip == 1 ? kombinacije(n, k, m) : tip == 2 ? kombinacijeSPonavljanjem(n, k, m) : tip == 3 ? variacije(n, k, m) : variacijeSPonavljanjem(n, k, m);
+		int[] result = tip == 1 ? kombinacije(n, k, m) : tip == 2 ? kombinacijeSPonavljanjem(n, k, m) : tip == 3 ? variacije(n, k, m) : variacijeSPonavljanjem(n, k, m);
 		s.close();
 		System.out.print("[" + result[0]);
 		for (int i = 1; i < result.length; i++) {
@@ -29,26 +29,26 @@ public class Tekm_63250060 {
 		return result;
 	}
 
-	public static BigInteger stKombinacij(byte n, byte k) {
-		return fakulteta(n).divide(fakulteta(k).multiply(fakulteta((byte) (n - k))));
+	public static BigInteger stKombinacij(int n, int k) {
+		return fakulteta(n).divide(fakulteta(k).multiply(fakulteta((int) (n - k))));
 	}
 
-	public static BigInteger stKombinacijSPonavljanjem(byte n, byte k) {
-		return fakulteta(n + k - 1).divide(fakulteta(k).multiply(fakulteta((byte) (n - 1))));
+	public static BigInteger stKombinacijSPonavljanjem(int n, int k) {
+		return fakulteta(n + k - 1).divide(fakulteta(k).multiply(fakulteta((int) (n - 1))));
 	}
 
 
-	public static byte[] kombinacije(byte n, byte k, BigInteger m) {
+	public static int[] kombinacije(int n, int k, BigInteger m) {
 		BigInteger vsota = BigInteger.ZERO;
-		byte[] result = new byte[k];
-		for (byte i = 0; i < k; i++) {
-			byte j;
+		int[] result = new int[k];
+		for (int i = 0; i < k; i++) {
+			int j;
 			if (i == 0)
 				j = 1;
 			else
-				j = (byte) (result[i - 1] + 1);
-			while (vsota.add(stKombinacij((byte) (n - j), (byte) (k - i - 1))).compareTo(m) < 0) {
-    			vsota = vsota.add(stKombinacij((byte) (n - j), (byte) (k - i - 1)));
+				j = (int) (result[i - 1] + 1);
+			while (vsota.add(stKombinacij((int) (n - j), (int) (k - i - 1))).compareTo(m) < 0) {
+    			vsota = vsota.add(stKombinacij((int) (n - j), (int) (k - i - 1)));
 				j++;
 			}
 			result[i] = j;
@@ -56,17 +56,17 @@ public class Tekm_63250060 {
 		return result;
 	}
 
-	public static byte[] kombinacijeSPonavljanjem(byte n, byte k, BigInteger m) {
+	public static int[] kombinacijeSPonavljanjem(int n, int k, BigInteger m) {
 		BigInteger vsota = BigInteger.ZERO;
-		byte[] result = new byte[k];
-		for (byte i = 0; i < k; i++) {
-			byte j;
+		int[] result = new int[k];
+		for (int i = 0; i < k; i++) {
+			int j;
 			if (i == 0)
 				j = 1;
 			else
-				j = (byte) (result[i - 1]);
-			while (vsota.add(stKombinacijSPonavljanjem((byte) (n - j + 1), (byte) (k - i - 1))).compareTo(m) < 0) {
-				vsota = vsota.add(stKombinacijSPonavljanjem((byte) (n - j + 1), (byte) (k - i - 1)));
+				j = (int) (result[i - 1]);
+			while (vsota.add(stKombinacijSPonavljanjem((int) (n - j + 1), (int) (k - i - 1))).compareTo(m) < 0) {
+				vsota = vsota.add(stKombinacijSPonavljanjem((int) (n - j + 1), (int) (k - i - 1)));
 				j++;
 			}
 			result[i] = j;
@@ -74,34 +74,34 @@ public class Tekm_63250060 {
 		return result;
 	}
 
-	public static byte[] variacije(byte n, byte k, BigInteger m) {
-		byte[] result = new byte[k];
+	public static int[] variacije(int n, int k, BigInteger m) {
+		int[] result = new int[k];
 		BigInteger mod = m;
-		for (byte i = 0; i < k; i++) {
+		for (int i = 0; i < k; i++) {
 			BigInteger[] deljenjeInMod = mod.divideAndRemainder(fakulteta(n - i - 1).divide(fakulteta(n - k)));
 			deljenjeInMod[0] = deljenjeInMod[1].compareTo(BigInteger.ZERO) == 0 ? deljenjeInMod[0] : deljenjeInMod[0].add(BigInteger.ONE);
 			deljenjeInMod[1] = deljenjeInMod[1].compareTo(BigInteger.ZERO) == 0 ? fakulteta(n - i - 1).divide(fakulteta(n - k)) : deljenjeInMod[1];
-			result[i] = deljenjeInMod[0].byteValue();
+			result[i] = deljenjeInMod[0].intValue();
 			mod = deljenjeInMod[1];
 		}
 
 		for (int i = 0; i < k; i++) {
-			byte[] sorted = Arrays.copyOf(result, i);
+			int[] sorted = Arrays.copyOf(result, i);
 			Arrays.sort(sorted);
 			for (int j = 0; j < i; j++) {
-				result[i] = result[i] >= sorted[j] ? (byte) (result[i] + 1) : result[i];
+				result[i] = result[i] >= sorted[j] ? (int) (result[i] + 1) : result[i];
 			}
 
 		}
 		return result;
 	}
 
-	public static byte[] variacijeSPonavljanjem(byte n, byte k, BigInteger m) {
-		byte[] result = new byte[k];
-		for (byte i = 1; i <= k; i++) {
+	public static int[] variacijeSPonavljanjem(int n, int k, BigInteger m) {
+		int[] result = new int[k];
+		for (int i = 1; i <= k; i++) {
 			BigInteger[] deljenje = m.divideAndRemainder(BigInteger.valueOf(n).pow(k -i));
 			BigInteger ceil = deljenje[1].compareTo(BigInteger.ZERO) == 0 ? deljenje[0] : deljenje[0].add(BigInteger.ONE);
-			byte moduloZn = ceil.mod(BigInteger.valueOf(n)).byteValue();
+			int moduloZn = ceil.mod(BigInteger.valueOf(n)).intValue();
 			result[i - 1] = moduloZn == 0 ? n : moduloZn;
 		}
 		return result;
